# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Task Management application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "BranchRequest": {
        "type": "object",
        "required": [
          "name",
          "address",
          "phone"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "description": "Name of the branch",
            "example": "Toronto Branch"
          },
          "address": {
            "type": "string",
            "minLength": 3,
            "description": "Physical address of the branch",
            "example": "440 Queen St W, Toronto, ON, M5V 2A8"
          },
          "phone": {
            "type": "string",
            "minLength": 3,
            "description": "Phone number of the branch",
            "example": "+1-416-980-2500"
          }
        }
      },
      "Branch": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BranchRequest"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Unique identifier for the branch",
                "example": "branch_456def"
              }
            },
            "required": [
              "id"
            ]
          }
        ]
      },
      "ValidationError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or code",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Validation failed"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "issue": {
                  "type": "string",
                  "example": "Employee email must be a valid email"
                }
              }
            },
            "description": "Detailed validation errors"
          }
        }
      },
      "EmployeeRequest": {
        "type": "object",
        "required": [
          "name",
          "position",
          "department",
          "email",
          "phone",
          "branchId"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "description": "Full name of the employee",
            "example": "John Smith"
          },
          "position": {
            "type": "string",
            "minLength": 3,
            "description": "Job position of the employee",
            "example": "Senior Developer"
          },
          "department": {
            "type": "string",
            "minLength": 3,
            "description": "Department of the employee",
            "example": "Engineering"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the employee",
            "example": "john.smith@company.com"
          },
          "phone": {
            "type": "string",
            "minLength": 3,
            "description": "Phone number of the employee",
            "example": "+1-555-0123"
          },
          "branchId": {
            "type": "number",
            "description": "ID of the branch the employee belongs to",
            "example": 1
          }
        }
      },
      "Employee": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EmployeeRequest"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Unique identifier for the employee",
                "example": "emp_123abc"
              }
            },
            "required": [
              "id"
            ]
          }
        ]
      },
      "ApiResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "description": "Status of the API response",
            "example": "success"
          },
          "data": {
            "description": "Response data (employee object or array)"
          },
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Employee created successfully"
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/branches": {
      "post": {
        "tags": [
          "Branches"
        ],
        "summary": "Create a new branch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Branch created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      },
      "get": {
        "tags": [
          "Branches"
        ],
        "summary": "Retrieve a list of branches",
        "responses": {
          "200": {
            "description": "A JSON array of branch objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/branches/{id}": {
      "get": {
        "tags": [
          "Branches"
        ],
        "summary": "Get a branch by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the branch"
          }
        ],
        "responses": {
          "200": {
            "description": "Branch object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "404": {
            "description": "Branch not found"
          }
        }
      },
      "put": {
        "tags": [
          "Branches"
        ],
        "summary": "Update a branch by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the branch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branch updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "Branch not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Branches"
        ],
        "summary": "Delete a branch by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the branch"
          }
        ],
        "responses": {
          "200": {
            "description": "Branch deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Branch not found"
          }
        }
      }
    },
    "/employees": {
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Create a new employee",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      },
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Retrieve a list of all employees",
        "responses": {
          "200": {
            "description": "A JSON array of employee objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/employees/{id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get an employee by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the employee"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found"
          }
        }
      },
      "put": {
        "tags": [
          "Employees"
        ],
        "summary": "Update an employee by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the employee"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "Employee not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Employees"
        ],
        "summary": "Delete an employee by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique identifier of the employee"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found"
          }
        }
      }
    },
    "/employees/branch/{branchId}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get all employees for a specific branch",
        "parameters": [
          {
            "in": "path",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The branch ID to filter employees by"
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON array of employees in the branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Branch not found"
          }
        }
      }
    }
  },
  "tags": []
}